requirements.txt
gitPython
matplotlib